Stop using seamless VPN handover technique when restarting the tunnel
There are reports of buggy behaviour using this technique on certain
phone makes and OS combinations, namely, the traffic is no longer routed
through the tunnel after the restart. There are similar reports by other
OS projects on the internet, see
https://issuetracker.google.com/issues/76356827
https://github.com/AdguardTeam/AdguardForAndroid/issues/1770

The solution is to try and restart the underlying Psiphon tunnel without
bringing down the VPN interface whenever possible when settings change
requires a restart, and do a complete VPN service teardown and restart
when unavoidable, for example, when the VPN exclusion settings change.